route-service/requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
pydantic
fastapi
sqlalchemy
psycopg2
python-jose[cryptography]
python-multipart
